requirements.txt
farm-haystack>=1.15.0
fastapi>=0.95.2
uvicorn>=0.22.0
sentence-transformers>=2.2.2
qdrant-client>=1.7.0
python-dotenv
pypdf
python-docx
google-generativeai
numpy
scikit-learn
pydantic
python-multipart
